Follow up questions per story appear to the user
